Refactor tasks into there own modules. Add Changelog to archive.
